(svn r8055) -Codechange: Replace the different max, dmax, maxu whatever macros by a simple template function max(), that requires two arguments of the same type. While I'm at it change a variable called "max" to "maxval" in a function that calls max().

+++ src/main_gui.cpp
@@ -1054,7 +1054,7 @@ void ZoomInOrOutToCursorWindow(bool in, Window *w)
 static void CommonRaiseLowerBigLand(TileIndex tile, int mode)
 {
 	int sizex, sizey;
-	byte h;
+	uint h;
 
 	_generating_world = true; // used to create green terraformed land
